requirements.txt
aio_pika==9.5.5
fastapi==0.116.0
langchain==0.3.26
langchain_anthropic==0.3.17
langchain_community==0.3.27
langchain_core==0.3.68
langchain_google_genai==2.1.7
langchain_groq==0.3.5
newspaper3k==0.2.8
pika==1.3.2
pydantic==2.11.7
python-dotenv==1.1.1
redis==6.2.0
uvicorn==0.35.0
langchain_groq==0.3.5
uvicorn[standard]
lxml
lxml_html_clean
transformers>=4.0.0
httpx